hummingbird-server/src/test/java/com/vaadin/hummingbird/StateNodeTest.java
package com.vaadin.hummingbird;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.Collections;
import java.util.List;
import java.util.function.Consumer;

import org.junit.Assert;
import org.junit.Test;

import com.vaadin.hummingbird.change.NodeAttachChange;
import com.vaadin.hummingbird.change.NodeChange;
import com.vaadin.hummingbird.change.NodeDetachChange;
import com.vaadin.hummingbird.namespace.ElementDataNamespace;
import com.vaadin.hummingbird.namespace.ElementPropertiesNamespace;

public class StateNodeTest {
    @Test
    public void newNodeState() {
        StateNode node = createEmptyNode();

        NodeOwner owner = node.getOwner();

        Assert.assertNotNull("New node should have an owner", owner);

        Assert.assertEquals("New node shold have unassigned id", -1,
                node.getId());

        Assert.assertTrue("Owner should know about the new node",
                owner.getNodes().contains(node));

        Assert.assertFalse("Node should not be attached", node.isAttached());
    }

    @Test
    public void nodeContainsDefinedNamespaces() {
        StateNode node = new StateNode(
                Arrays.asList(ElementDataNamespace.class));

        ElementDataNamespace namespace = node
                .getNamespace(ElementDataNamespace.class);

        Assert.assertNotNull("Should have namespace defined in constructor",
                namespace);

        ElementPropertiesNamespace missingNamespace = node
                .getNamespace(ElementPropertiesNamespace.class);

        Assert.assertNull(
                "Should not have namespace that wasn't defined in constructor",
                missingNamespace);
    }

    @Test
    public void testAttachDetachChangeCollection() {
        StateNode node = createEmptyNode();

        List<NodeChange> changes = new ArrayList<>();
        Consumer<NodeChange> collector = changes::add;

        node.collectChanges(collector);

        Assert.assertTrue("Node should have no changes", changes.isEmpty());

        // Attach node
        node.setParent(new StateTree().getRootNode());

        node.collectChanges(collector);

        Assert.assertEquals("Should have 1 change", 1, changes.size());
        Assert.assertTrue("Should have attach change",
                changes.get(0) instanceof NodeAttachChange);
        changes.clear();

        node.collectChanges(collector);
        Assert.assertTrue("Node should have no changes", changes.isEmpty());

        // Detach node
        node.setParent(null);

        node.collectChanges(collector);
        Assert.assertEquals("Should have 1 change", 1, changes.size());
        Assert.assertTrue("Should have detach change",
                changes.get(0) instanceof NodeDetachChange);
        changes.clear();
    }

    public static StateNode createEmptyNode() {
        return createEmptyNode("Empty node");
    }

    public static StateNode createEmptyNode(String toString) {
        return new StateNode(Collections.emptyList()) {
            @Override
            public String toString() {
                return toString;
            }
        };
    }
}
